+++ src/Helpdesk/GetHelpdeskV1TicketsByTicketIdResponseDto.cs
@@ -4,7 +4,7 @@
 // Created          : 2024-06-24
 //
 // Last Modified By : yxr
-// Last Modified On : 2024-06-24
+// Last Modified On : 2024-06-28
 // ************************************************************************
 // <copyright file="GetHelpdeskV1TicketsByTicketIdResponseDto.cs" company="Vicente Yu">
 //     MIT
@@ -146,7 +146,7 @@ public record GetHelpdeskV1TicketsByTicketIdResponseDto
             /// <para>示例值：1690970837624</para>
             /// </summary>
             [JsonPropertyName("created_at")]
-            public int? CreatedAt { get; set; }
+            public long? CreatedAt { get; set; }
 
             /// <summary>
             /// <para>备注ID</para>
@@ -252,7 +252,7 @@ public record GetHelpdeskV1TicketsByTicketIdResponseDto
             /// <summary>
             /// <para>英文描述</para>
             /// <para>必填：否</para>
-            /// <para>示例值：Idon'tunderstand</para>
+            /// <para>示例值：I don't understand</para>
             /// </summary>
             [JsonPropertyName("en_us")]
             public string? EnUs { get; set; }
@@ -347,7 +347,7 @@ public record GetHelpdeskV1TicketsByTicketIdResponseDto
             /// <summary>
             /// <para>展示名称</para>
             /// <para>必填：否</para>
-            /// <para>示例值：displayname</para>
+            /// <para>示例值：display name</para>
             /// </summary>
             [JsonPropertyName("display_name")]
             public string? DisplayName { get; set; }
